+++ pkg/azureupdate/error.go
@@ -12,3 +12,12 @@ var invalidConfigError = &microerror.Error{
 func IsInvalidConfig(err error) bool {
 	return microerror.Cause(err) == invalidConfigError
 }
+
+var masterCIDRChangeError = &microerror.Error{
+	Kind: "masterCIDRChangeError",
+}
+
+// IsMasterCIDRChange asserts masterCIDRChangeError.
+func IsMasterCIDRChange(err error) bool {
+	return microerror.Cause(err) == masterCIDRChangeError
+}